# apps/backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Koyomi Gal 后端 API",
        "title": "Koyomi Gal API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
...
        "/api/v1/comments/{id}/like": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "当前用户点赞评论，原子更新点赞计数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "点赞评论",
                "operationId": "likeComment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "评论 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "点赞结果",
                        "schema": {
                            "$ref": "#/definitions/dto.CommentLikeDataResponse"
                        }
                    },
                    "400": {
                        "description": "评论 ID 格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户登录失效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "评论不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "已点赞该评论",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "点赞失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "当前用户取消点赞评论，原子更新点赞计数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "comments"
                ],
                "summary": "取消点赞评论",
                "operationId": "unlikeComment",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "评论 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "取消结果",
                        "schema": {
                            "$ref": "#/definitions/dto.CommentLikeDataResponse"
                        }
                    },
                    "400": {
                        "description": "评论 ID 格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户登录失效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "评论不存在或未点赞",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "取消点赞失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/posts/{id}/favorite": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "当前用户收藏帖子，原子更新收藏计数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "收藏帖子",
                "operationId": "favoritePost",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "帖子 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "收藏结果",
                        "schema": {
                            "$ref": "#/definitions/dto.PostFavoriteDataResponse"
                        }
                    },
                    "400": {
                        "description": "帖子 ID 格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户登录失效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "帖子不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "已收藏该帖子",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "收藏失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "当前用户取消收藏帖子，原子更新收藏计数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "取消收藏帖子",
                "operationId": "unfavoritePost",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "帖子 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "取消结果",
                        "schema": {
                            "$ref": "#/definitions/dto.PostFavoriteDataResponse"
                        }
                    },
                    "400": {
                        "description": "帖子 ID 格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户登录失效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "帖子不存在或未收藏",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "取消收藏失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/posts/{id}/like": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "当前用户点赞帖子，原子更新点赞计数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "点赞帖子",
                "operationId": "likePost",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "帖子 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "点赞结果",
                        "schema": {
                            "$ref": "#/definitions/dto.PostLikeDataResponse"
                        }
                    },
                    "400": {
                        "description": "帖子 ID 格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户登录失效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "帖子不存在",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "409": {
                        "description": "已点赞该帖子",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "点赞失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "当前用户取消点赞帖子，原子更新点赞计数",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "posts"
                ],
                "summary": "取消点赞帖子",
                "operationId": "unlikePost",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "帖子 ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "取消结果",
                        "schema": {
                            "$ref": "#/definitions/dto.PostLikeDataResponse"
                        }
                    },
                    "400": {
                        "description": "帖子 ID 格式不正确",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "用户登录失效",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "帖子不存在或未点赞",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "取消点赞失败",
                        "schema": {
                            "$ref": "#/definitions/response.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "dto.CommentLikeData": {
            "type": "object",
            "properties": {
                "like_count": {
                    "type": "integer",
                    "example": 6
                },
                "liked": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.CommentLikeDataResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/dto.CommentLikeData"
                },
                "msg": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "dto.PostFavoriteData": {
            "type": "object",
            "properties": {
                "favorite_count": {
                    "type": "integer",
                    "example": 3
                },
                "favorited": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.PostFavoriteDataResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/dto.PostFavoriteData"
                },
                "msg": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
        "dto.PostLikeData": {
            "type": "object",
            "properties": {
                "like_count": {
                    "type": "integer",
                    "example": 11
                },
                "liked": {
                    "type": "boolean",
                    "example": true
                }
            }
        },
        "dto.PostLikeDataResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 0
                },
                "data": {
                    "$ref": "#/definitions/dto.PostLikeData"
                },
                "msg": {
                    "type": "string",
                    "example": "success"
                }
            }
        },
...
        "response.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 233
                },
                "msg": {
                    "type": "string",
                    "example": "业务错误"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "传入 \"Bearer {access_token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}